refactor(tests): migrate logged-in sidebar #507
