+++ routes/admin/authenticated/sidebar/community-settings/recipient/page.spec.js
