+++ routes/admin/authenticated/sidebar/mobilizations-settings-analytics/page.spec.js
