+++ routes/admin/authenticated/sidebar/templates-choose-global/page.spec.js
