+++ routes/admin/authenticated/sidebar/templates-choose/page.spec.js
