+++ routes/admin/authenticated/sidebar/templates-list/page.spec.js
