+++ routes/admin/authenticated/sidebar/widgets-donation-settings/finish/page.spec.js
