+++ routes/admin/authenticated/sidebar/widgets-form-settings/autofire/page.spec.js
