+++ routes/admin/authenticated/sidebar/widgets-form-settings/finish/page.spec.js
