+++ routes/admin/authenticated/sidebar/widgets-pressure-settings/autofire/page.spec.js
